remove most floating point math and disable floating point unit (#391)
* remove most floating point math and disable floating point unit

* respond to Chase's comments

* changed lookup table to const

* remove floats in test case

+++ Apps/Inc/Temperature.h
@@ -45,7 +45,7 @@ ErrorStatus Temperature_ChannelConfig(uint8_t tempChannel);
  * @param mV from ADC
  * @return temperature in Celsius (Fixed Point with .001 resolution) 
  */
-int milliVoltToCelsius(float milliVolt);
+int milliVoltToCelsius(uint32_t milliVolt);
 
 /** Temperature_UpdateSingleChannel
  * Stores and updates the new measurements received on one particular temperature sensor
